+++ NEWS.md
@@ -8,6 +8,7 @@
 - modify getmass to accept negative value
 - add getms1anno for MS1 annotation by local compounds database with predefined paired mass distances
 - add support to save target data csv without change default name
+- add retention time alignment function for peaks lists
 
 # enviGCMS 0.6.6